Pull cap phase analysis out of JakeMan
